requirements.txt
# Core dependencies
numpy>=1.24.0
scikit-learn>=1.3.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0

# Text embedding
sentence-transformers>=2.2.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: for better performance
faiss-cpu>=1.7.0  # For faster similarity search 